+++ src/backend/services/file.py
@@ -4,7 +4,6 @@ import pandas as pd
 from docx import Document
 from fastapi import HTTPException
 from fastapi import UploadFile as FastAPIUploadFile
-from pypdf import PdfReader
 from python_calamine.pandas import pandas_monkeypatch
 
 import backend.crud.conversation as conversation_crud
@@ -16,6 +15,7 @@ from backend.database_models.conversation import ConversationFileAssociation
 from backend.database_models.database import DBSessionDep
 from backend.database_models.file import File as FileModel
 from backend.schemas.file import File, UpdateFileRequest
+from backend.services import utils
 
 MAX_FILE_SIZE = 20_000_000  # 20MB
 MAX_TOTAL_FILE_SIZE = 1_000_000_000  # 1GB
@@ -352,7 +352,7 @@ async def get_file_content(file: FastAPIUploadFile) -> str:
     file_extension = get_file_extension(file.filename)
 
     if file_extension == PDF_EXTENSION:
-        return read_pdf(file_contents)
+        return utils.read_pdf(file_contents)
     elif file_extension == DOCX_EXTENSION:
         return read_docx(file_contents)
     elif file_extension in [
@@ -366,26 +366,6 @@ async def get_file_content(file: FastAPIUploadFile) -> str:
         return read_excel(file_contents)
 
     raise ValueError(f"File extension {file_extension} is not supported")
-
-
-def read_pdf(file_contents: bytes) -> str:
-    """Reads the text from a PDF file using PyPDF2
-
-    Args:
-        file_contents (bytes): The file contents
-
-    Returns:
-        str: The text extracted from the PDF
-    """
-    pdf_reader = PdfReader(io.BytesIO(file_contents))
-    text = ""
-
-    # Extract text from each page
-    for page in pdf_reader.pages:
-        page_text = page.extract_text()
-        text += page_text
-
-    return text
 
 
 def read_excel(file_contents: bytes) -> str:
